1point2/login-module/requirements.txt
Flask>=2.0.0
mysql-connector-python>=8.0.0
bcrypt>=4.0.0